Произведена модуляция
